// angular-app/SeaDuel/src/swagger/model/opponentBoard.ts
/**
 * SeaDuelServer
 * SeaDuelServer
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface OpponentBoard {
  width?: number;
  height?: number;
  boardData?: Array<{
    type: "Unknown" | "Hit" | "Miss" | "Boat";
  }>;
}